+++ activejob/test/cases/argument_serialization_test.rb
@@ -9,6 +9,7 @@ require "active_support/core_ext/integer/time"
 require "active_support/duration"
 require "jobs/kwargs_job"
 require "jobs/arguments_round_trip_job"
+require "jobs/association_loading_job"
 require "support/stubs/strong_parameters"
 
 class ArgumentSerializationTest < ActiveSupport::TestCase
@@ -44,6 +45,7 @@ class ArgumentSerializationTest < ActiveSupport::TestCase
   end
 
   setup do
+    AssociationLoadingJob.locators.clear
     @person = Person.find("5")
     @original_serializers = ActiveJob::Serializers.serializers
   end
@@ -266,17 +268,35 @@ class ArgumentSerializationTest < ActiveSupport::TestCase
     assert_match "Unable to serialize Person without an id.", err.message
   end
 
+  if ENV["AJ_INTEGRATION_TESTS"]
+    test "should locate records compatible with strict loading locator" do
+      AssociationLoadingJob.locator_options "Article", includes: :tags
+
+      article = Article.create!(tags_attributes: [{ name: "a" }, { name: "b" }])
+
+      assert_arguments_roundtrip [article, [:tags]], job_class: AssociationLoadingJob
+    end
+
+    test "should raise a strict loading error when locating records without locator_options configuration" do
+      article = Article.create!(tags_attributes: [{ name: "a" }, { name: "b" }])
+
+      assert_raises ActiveRecord::StrictLoadingViolationError, match: "`Article` is marked for strict_loading" do
+        AssociationLoadingJob.perform_later(article, [:tags])
+      end
+    end
+  end
+
   private
     def assert_arguments_unchanged(*args)
       assert_arguments_roundtrip args
     end
 
-    def assert_arguments_roundtrip(args)
-      assert_equal args, perform_round_trip(args)
+    def assert_arguments_roundtrip(args, **opts)
+      assert_equal args, perform_round_trip(args, **opts)
     end
 
-    def perform_round_trip(args)
-      ArgumentsRoundTripJob.perform_later(*args) # Actually performed inline
+    def perform_round_trip(args, job_class: ArgumentsRoundTripJob)
+      job_class.perform_later(*args) # Actually performed inline
 
       JobBuffer.last_value
     end