+++ activejob/test/jobs/association_loading_job.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class AssociationLoadingJob < ArgumentsRoundTripJob
+  def perform(record, associations = [])
+    associations.each { |association| record.send(association).load }
+
+    super
+  end
+end